+++ libs/core/src/lib/bar/public_api.ts
@@ -7,3 +7,4 @@ export * from './directives/bar-right.directive';
 export * from './button-bar/button-bar.component';
 export * from './deprecated-bar-content-density.directive';
 export * from './deprecated-bar-button-content-density.directive';
+export * from './tokens';