+++ libs/core/src/lib/flexible-column-layout/tokens.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core';
+
+export const FD_FLEXIBLE_COLUMN_LAYOUT_COMPONENT = new InjectionToken('FdFlexibleColumnLayoutComponent');