+++ libs/core/src/lib/list/list-navigation-item/list-navigation-item.component.ts
@@ -9,7 +9,7 @@ import {
     Input,
     Optional
 } from '@angular/core';
-import { IconComponent } from '@fundamental-ngx/core/icon';
+import { FD_ICON_COMPONENT, IconComponent } from '@fundamental-ngx/core/icon';
 import { KeyUtil, RtlService } from '@fundamental-ngx/cdk/utils';
 import { ENTER, LEFT_ARROW, RIGHT_ARROW, SPACE } from '@angular/cdk/keycodes';
 import { FocusableOption } from '@angular/cdk/a11y';
@@ -61,7 +61,7 @@ export class ListNavigationItemComponent implements AfterContentInit, AfterViewI
     _listNavigationItemArrow: ListNavigationItemArrowDirective;
 
     /** @hidden */
-    @ContentChild(IconComponent)
+    @ContentChild(FD_ICON_COMPONENT)
     _iconComponent: IconComponent;
 
     /** @hidden */