+++ libs/core/src/lib/radio/public_api.ts
@@ -1,3 +1,4 @@
 export * from './radio.module';
 export * from './deprecated-radio-button-compact.directive';
 export * from './radio-button/radio-button.component';
+export * from './tokens';